select box (detail view), language strings, etc....
added language strings, selectbox detailview. The specific detailview
was not working as the query was wrong. It contained the catsid field
and that was removed in v1.1... so there have been some alterations.
like button in detailview, setting is in config.xml for now, is easier
then altering the tables.