deleted the assets folder + altered colspan in default_table
The assets folder is deleted as it is not being used anymore, the files
are now located in the media folder. There are references within the
help, but that has to be updated in general. The colspan of the
default_table has been altered to show "no events" centered in frontend