Improved mailer plugin & email by category
Now, JEM can send email notification to several case, in function of
action (new event, update event, delete event, new venue, update venue)
and select type user (owner event, user group, email category,
distribution list of plugin):

user group: The users of the group associated with the category of the
event.
email category: A mailing list can be assigned to the category to notify
new events to a category. In the future this parameter will select list
of emails or name from the list of Acymailing (plugin).
It can notify to registered users of changes in the event (any update,
include move to trash (delete).

I have also ready spanish language file for this new plugin.
I hope you like the changes in this new version of the plugin.