Review recurrence process (alpha 1) #1762
Added new option of  recurrence type: year.
When edited, show the options saved in the fields (back and front).
New shield by days, week, month, year in the global setting.
Now all events of recurrence are generated and saved.
In the event view (legacy and responsive), there is a new icon to edit the root event. (first event).