Review recurrence process (beta 1) #1762
Include the new field 'seriesbooking' into events table. This option is set in the extended tab of event. If it's active, the user books places in all events of the series (only event with recurrence active) from now until end date of recurrence. The places are booked if the user doesn't register and there are any places free. Unregistered is by event.