Important: Fix the update JEM schema in database #1816
When using ALTER TABLE, Joomla requires the COLUMN keyword in the array_string[4] if it wants to add a new column to the table. This is [defined in \libraries\src\Schema\ChangeItem\MysqlChangeItem.php [line 90 in JL 5.1.2].
For this reason, these columns aren't created during the migration process or when JoomlWhen using ALTER TABLE, Joomla requires the COLUMN keyword in the array_string[4] if it needs to add a new column to the table. This is defined in \libraries\src\Schema\ChangeItem\MysqlChangeItem.php [line 90 in Joomla 5.1.2]. For this reason, these columns weren't created during the migration process or when Joomla attempted to fix the JEM schema. However, with the recent commit, this issue is now resolved.a try to fix the JEM squema, these columns weren't created. Now with commit, works.